Shadows no longer behaving wonky
The key was adding `sun.autoUpdateExtends = false`

This tells the light source (our directional light) to not recompute shadow position. I'm not sure why, but recomputing shadow position seems to result in a negative feedback loop that slowly mutates the shadow map to the point that it's not useful.

This commit also adds self-shadowing to the reference cubes

closes #1