Smooth collisions
Ok, this commit doesn't _exactly_ smooth collisions. What it does instead is simplify the geometry of the player mesh. The player mesh is now a box instead of a cylindar. The box geometry does away with the strange movement behavior of the cylindar mesh; there are only four sides to collide with, so stange subtle movements are not possible as there aren't faces at weird angles to collide with anything.

closes #7