Use Babylon's fullscreen api and the browser api to manage fullscreen
Babylon has two methods which we use two manage fullscreen:

* `engine.switchFullscreen(lockPointer: bool)`
* `engine.setSize(width: int, height: int)`

Calling `engine.switchFullscreen(true)` switches the browser to fullscreen mode, but it does **not** update the size of the `canvas`.

To ensure our `canvas` fill the screen we must manually set it's size using `engine.setSize(screen.width, screen.height)`. The fact that we're using `screen.width` and not `window.innerWidth` is important. When we go into fullscreen mode we do not want to resize our `canvas` to match the browser window's size, we want it to match the screen's size. Using `screen` means that a player can have their browser at any size they want - when they enter fullscreen mode, the game `canvas` will fill their screen.

Also note that we call `canvas.focus()`. This is critically important because without this call the player would have to way to move their character. For our game to receive keyboard input the `canvas` element must be focused. Since our `engine.switchFullscreen` call passes `true` as it's argument the mouse pointer has disappeared - meaning there is no way for the player to click on the `canvas` to make it focused.

Expecting a player to click on the `canvas` to play the game would be extremely poor design as we'd be expecting unintuitive behavior of our player. By making sure the `canvas` is automatically focused when appropriate we present a fully playable game to our player when they enter fullscreen mode.

Also notice how we're handling entering and exiting fullscreen mode.

We explicitly handle the keyboard input to enter fullscreen by checking for an "Enter/Return" (keyCode 13) `keydown` event.

Exiting fullscreen mode is handled differently. The browser has built-in support for handling `fullscreenchange` events. For some god-forsaken reason these events all still require vendor prefixes. So, the event I'm checking for now is `onwebkitfullscreenchange` because I'm using Chrome. The `onwebkitfullscreenchange` event is fired when fullscreen mode is entered and exited. Unfortunately, as far as I can tell, the event data for both cases is exactly the same - meaning we need a flag external to the event we can use to track whether we're entering or exiting fullscreen mode.

Happily, Babylon has us covered. There is a `engine.isFullscreen` property that tells us whether we're currently in fullscreen mode or not.

We check this property in the `onwebkitfullscreenchange` event handler. The browser handles the `esc` key press (which exits fullscreen mode) for us, so by the time we get to this event handler the browser has already done what's needed to remove us from fullscreen mode (which includes unlocking the pointer). Checking the `engine.isFullscreen` in this handler will return `false`. `false` means that we need to manually resize out `canvas` just like we did when entering fullscreen mode.

To resize our `canvas` we just use the width and height values we saved previously in the `original_canvas_width` and `original_canvas_height` properties.

With all that in place we have basic, funcional support for entering and exiting fullscreen mode!