Wipe away the old old movement code
In order to update movement to take rotation into account, I needed to do a complete rewrite of the movement code. Not only does this rewrite close an issue, it also drastically cleans up the `movement.js` code as well as it's interface, which `index.js` uses.

This commit purges all the old code so that the next commit, where all the cool cleaned up stuff is added, is easier to understand and clearer to read.