Ingore device's pixelRatio property
Passing `false` for the fourth parameter to the `Engine` constructor tells Babylon to ignore a device's pixelRatio.

Many modern dislays are what's called High DPI. I tried searching for a good definition of High DPI but couldn't find one, so I'll provide it here:

High DPI stands for High Dots Per Inch. Basically it means that there are a crap ton of pixels in a single inch of screen space.

What many modern devices have started doing is presenting their resolutions relative to their `pixelRatio`. So, for instance, my laptop displays a resolution of 1440 x 900 is the settings panel. This is not the actual resolution of my laptop monitor. The actual resolution of my monitor is found by taking the numbers reported via settings and multiplying them by the `pixelRatio`. My monitor has a `pixelPatio` of 2. To get the actual number of pixel my monitor has, I need to multiply both 1440 and 900 by 2. This gives the actual, really really real, resolution of 2880 x 1800.

This whole idea may seem very strange and unnecessarily complex, but it's actually a wonderful thing: by reporting the resolution of 1440 x 900, my monitor gives applications a choice regarding how they want to render things to the screen: using the reported resolution, or the actual resolution (which includes multiplying by the `pixelRatio`). This is great because apps that need crystal clear, razor sharp graphics (and render content by themselves per-pixel, like games) can choose to either render at the reported resolution and save serious performance, or render and the full, actuall resolution (which takes roughly 4 times the work, and runs roughly 4 times slower, for a `pixelRatio` of 2) and get super clear, crisp gameplay.

For this project I'm choosing to ignore the `pixelRatio`. This means that Babylon will only need to render 1,296,000 pixels instead of 5,184,000 pixels per frame.

My laptop is geting old, so this performance saving is a big deal for me.